+++ CHANGELOG.md
@@ -6,9 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
-- `be.oSet`, Set object assertion
-- `be.weakSet`, WeakSet object assertion
-- `be.map`, Map object assertion
+- es6 `be.oSet`, Set object assertion
+- es6 `be.weakSet`, WeakSet object assertion
+- es6 `be.map`, Map object assertion
+- es6 `be.mapIterator`, Map Iterator object assertion
 
 ## [1.4.1] - 2017-07-10